requirements.txt
asgiref==3.7.2
Django==4.2.7
django-browser-reload==1.12.1
django-ckeditor==6.7.0
django-js-asset==2.1.0
django-tailwind==3.6.0
gunicorn==21.2.0
mysqlclient==2.2.0
packaging==23.2
Pillow==10.1.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
